+++ tests/test_render.py
@@ -1,11 +1,11 @@
 from dataclasses import dataclass
-from typing import Any, Sequence
+from typing import Sequence, Tuple
 
 import pytest
+from viewdom.h import html
 from wired import ServiceRegistry
 from wired.dataclasses import injected, Context, factory, register_dataclass
 
-from viewdom.h import html
 from viewdom_wired import render, component, register_component
 
 
@@ -88,4 +88,22 @@ def test_wired_renderer_children(registry: ServiceRegistry):
     container = registry.create_container()
     expected = '<h1>Hello</h1><div>Child</div>'
     actual = render(html('''<{Heading}>Child<//>'''), container)
+    assert expected == actual
+
+
+def test_wired_renderer_generics(registry: ServiceRegistry):
+    """ Tolerate usage of PEP 484 generics on type hints """
+
+    @dataclass
+    class LocalHeading:
+        names: Tuple[str, ...] = ('Name 1',)
+
+        def __call__(self):
+            name_one = self.names[0]
+            return html('''<h1>{name_one}</h1>''')
+
+    container = registry.create_container()
+    register_component(registry, LocalHeading, LocalHeading)
+    expected = '<h1>Name 1</h1>'
+    actual = render(html('''<{LocalHeading}/>'''), container)
     assert expected == actual