+++ CHANGES.rst
@@ -6,6 +6,8 @@ Unreleased
 
 - The ``Component`` protocol now indicates it returns a ``VDOM``
 
+- Tolerate usage of generics such as `Optional[str]`
+
 0.1.0
 =====